cards-server-app/requirements.txt
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
django-health-check==3.17.0
django-import-export==3.0.1
django-pictures==1.0.1
django-rest-passwordreset==1.3.0
django-stdimage==6.0.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-yasg==1.21.4
gunicorn==20.1.0
Pillow==9.4.0
psycopg2==2.9.4
psycopg2-binary==2.9.4
rest-social-auth==8.0.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
whitenoise==6.2.0

